+++ libauc/metrics/__init__.py
@@ -1,1 +1,2 @@
+from .metricsk import *
 from .metrics import *
